+++ src/mn_menu.c
@@ -1316,7 +1316,7 @@ static void M_QuitResponse(int ch)
     {
         return;
     }
-    if (D_CheckEndDoom() &&           // play quit sound only if showing ENDOOM
+    if (quit_sound &&                 // [Nugget]
         (!netgame || demoplayback) && // killough 12/98
         !nosfxparm)                   // avoid delay if no sound card
     {