+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **_Message Flash_** setting
 - **Support for powerup-timer icons**
   - Replaced `hud_stats_icons` with `hud_allow_icons` [1]
+- **Quit Sound** setting, enabled by default
 
 ## Changes